cleanup: move filter tabs to a different class
