+++ src/FilterTabs.tsx
@@ -1,0 +1,29 @@
+import React, { useState } from "react";
+
+interface FilterTabsProps {
+  current: string;
+  onFilterChange: (filter: string) => void;
+}
+
+export function FilterTabs({
+  current,
+  onFilterChange,
+}: FilterTabsProps) {
+  const getTabStyle = (buttonType) => ({
+    fontWeight: current === buttonType ? "bold" : "normal",
+    backgroundColor: current === buttonType ? "#aaa" : "#1a1a1a",
+    padding: "0.5rem 1rem",
+    cursor: "pointer",
+  });
+
+  return (
+    <div className="subcard" style={{
+      display: "flex",
+      justifyContent: "space-evenly",
+    }}>
+      <button style={getTabStyle("all")} onClick={() => onFilterChange("all")}>All</button>
+      <button style={getTabStyle("active")} onClick={() => onFilterChange("active")}>Active</button>
+      <button style={getTabStyle("completed")} onClick={() => onFilterChange("completed")}>Completed</button>
+    </div>
+  );
+}
